refactor(public-rsvp): rename endpoint modules for naming consistency
Both modules are the same public (non-member, token) RSVP feature, split by
lifecycle stage. Rename to a parallel submit/manage pair so the shared
"public_rsvp" family is explicit and neither reads as "not public":

- _public_rsvp.py       -> _public_rsvp_submit.py   (create: POST .../rsvp/)
- _my_rsvps.py          -> _public_rsvp_manage.py   (manage: GET/POST/DELETE /my-rsvps/)
- test_my_rsvps.py      -> test_public_rsvp_manage.py

Router vars renamed to match (public_rsvp_submit_router / public_rsvp_manage_router).
URLs, function names, and behavior are unchanged; regenerated openapi schema +
frontend types pick up the new operationId prefixes.
