+++ backend/tests/test_public_rsvp_manage.py
